// src/api/openapi/api/offer.service.ts
export const OfferFieldEnum = {
  Numeric: "numeric",
  Alphanumeric: "alphanumeric",
  Calculation: "calculation",
  Dropdown: "dropdown",
  Offertext: "offertext"
} as const;
export type OfferFieldEnum = typeof OfferFieldEnum[keyof typeof OfferFieldEnum];

export interface OfferField {
  id: number;
  label: string;
  description: string;
  type: OfferFieldEnum;
  calculation?: string;
  visible: boolean;
  unitId?: number;
}

export interface OfferFieldCreate {
  type: OfferFieldEnum;
  label: string;
  description: string;
  calculation?: string;
  unitId?: number;
}

export interface OfferFieldUpdate {
  label?: string;
  description?: string;
  calculation?: string;
  unitId?: number;
}


export interface OfferElementField extends OfferField {
  defaultValue: string;
}

export interface OfferElementType {
  id: number;
  title: string;
  description: OfferField;
  offertext: OfferElementField;
  price: OfferElementField;
  fields: OfferField[];
}

export interface OfferElementTypeCreate {
  title: string;
  offertext: string;
  price: string;
  fields: number[];
}

export interface OfferElementTypeUpdate {
  title: string;
  offertext: string;
  price: string;
  fields: number[];
}

const OFFTEXT_FIELD: OfferField = {
  id: 1,
  label: "Angebotstext",
  description: "",
  type: OfferFieldEnum.Calculation,
  visible: false,
  calculation: ""
};
const PRICE_FIELD: OfferField = {
  id: 2,
  label: "Preisberechnung",
  description: "",
  type: OfferFieldEnum.Calculation,
  visible: false,
  calculation: ""
};
const DESC_FIELD: OfferField = {
  id: 3,
  label: "Beschreibung",
  description: "",
  type: OfferFieldEnum.Alphanumeric,
  visible: false
};


@Injectable({
  providedIn: "root"
})
export class OfferService extends BaseService {

  constructor(protected httpClient: HttpClient, @Optional() @Inject(BASE_PATH) basePath: string | string[], @Optional() configuration?: Configuration) {
    super(basePath, configuration);
  }

  private getFieldsInternal(): OfferField[] {
    return JSON.parse(localStorage.getItem("offer-fields") ?? "[]").concat([OFFTEXT_FIELD, PRICE_FIELD, DESC_FIELD]);
  }

  private storeFieldsInternal(elems: OfferField[]) {
    localStorage.setItem("offer-fields", JSON.stringify(elems));
  }

  public getFields(): Observable<OfferField[]> {
    return of(this.getFieldsInternal());
  }

  public getField(fieldId: number): Observable<OfferField | undefined> {
    const fields: OfferField[] = this.getFieldsInternal();
    return of(fields.find(f => f.id === fieldId));
  }

  public countFields(): Observable<number> {
    const fields: OfferField[] = this.getFieldsInternal();
    return of(fields.length);
  }

  public patchField(fieldId: number, fieldPatch: OfferFieldUpdate): Observable<OfferField | undefined> {
    const fields: OfferField[] = this.getFieldsInternal();
    const f = fields.find(f => f.id === fieldId);
    if (f) {
      this.storeFieldsInternal(fields.map(f => f.id === fieldId ? {
        ...f,
        description: fieldPatch.description,
        unitId: fieldPatch.unitId,
        label: fieldPatch.label,
        calculation: fieldPatch.calculation
      } : f));
      return this.getField(fieldId);
    }
    return of(undefined);
  }

  public createField(fieldCreate: OfferFieldCreate): Observable<OfferField | undefined> {
    const fields: OfferField[] = this.getFieldsInternal();
    const f: OfferField = {
      id: Math.max(...fields.map(f => f.id)) + 1,
      label: fieldCreate.label,
      calculation: fieldCreate.calculation,
      unitId: fieldCreate.unitId,
      type: fieldCreate.type,
      description: fieldCreate.description,
      visible: true
    };
    this.storeFieldsInternal(fields.concat([f]));
    return of(f);
  }

  private getElementTypesInternal(): OfferElementType[] {
    return JSON.parse(localStorage.getItem("offer-element-types") ?? "[]");
  }

  private storeElementTypesInternal(elems: OfferElementType[]) {
    localStorage.setItem("offer-element-types", JSON.stringify(elems));
  }

  public getElementTypes(): Observable<OfferElementType[]> {
    return of(this.getElementTypesInternal());
  }

  public getElementType(fieldId: number): Observable<OfferElementType | undefined> {
    const fields: OfferElementType[] = this.getElementTypesInternal();
    return of(fields.find(f => f.id === fieldId));
  }

  public countElementTypes(): Observable<number> {
    const fields: OfferElementType[] = this.getElementTypesInternal();
    return of(fields.length);
  }

  public patchElementType(elemId: number, elementTypePatch: OfferElementTypeUpdate): Observable<OfferElementType | undefined> {
    const elementTypes = this.getElementTypesInternal();
    const elem = elementTypes.find(f => f.id === elemId);
    const fields = this.getFieldsInternal();
    this.storeElementTypesInternal(elementTypes.map(f => f.id === elemId ? {
      ...f,
      title: elementTypePatch.title,
      offertext: {
        ...f.offertext,
        defaultValue: elementTypePatch.offertext
      },
      price: {
        ...f.price,
        defaultValue: elementTypePatch.price
      },
      fields: fields.filter(f => elementTypePatch.fields.includes(f.id))
    } : f));
    return this.getElementType(elemId);
  }

  public createElementType(elementTypeCreate: OfferElementTypeCreate): Observable<OfferElementType | undefined> {
    const elementTypes = this.getElementTypesInternal();
    const fields = this.getFieldsInternal();
    const elem: OfferElementType = {
      id: Math.max(...elementTypes.map(f => f.id)) + 1,
      title: elementTypeCreate.title,
      description: DESC_FIELD,
      price: {
        ...PRICE_FIELD,
        defaultValue: elementTypeCreate.price
      },
      offertext: {
        ...OFFTEXT_FIELD,
        defaultValue: elementTypeCreate.offertext
      },
      fields: fields.filter(f => elementTypeCreate.fields.includes(f.id))
    };
    this.storeElementTypesInternal(elementTypes.concat([elem]));
    return of(elem);
  }

}